+++ webviz_subsurface/plugins/_bhp_qc/__init__.py
@@ -1,0 +1,2 @@
+from ._plugin import BhpQc
+from ._plugin_ids import PluginIds
