+++ webviz_subsurface/plugins/_bhp_qc/_error.py
@@ -1,0 +1,5 @@
+from dash import html
+
+
+def error(error_message: str) -> html.Div:
+    return html.Div(children=error_message, style={"color": "red"})
