+++ webviz_subsurface/plugins/_bhp_qc/_plugin_ids.py
@@ -1,0 +1,26 @@
+class PluginIds:
+    # pylint: disable=too-few-public-methods
+    class Stores:
+        # pylint: disable=too-few-public-methods
+        SELECTED_ENSEMBLE = "selected-ensemble"
+        SELECTED_WELLS = "selected-wells"
+        SELECTED_MAX_NUMBER_OF_WELLS = "selected-max-number-of-wells"
+        SELECTED_SORT_BY = "selected-sort-by"
+        SELECTED_ASCENDING_DESCENDING = "selected-ascending-descending"
+        SELECTED_PLOT_TYPE = "selected-plot-type"
+        SELECTED_STATISTICS = "selected-statistics"
+
+    class SharedSettings:
+        # pylint: disable=too-few-public-methods
+        FILTER = "filter"
+        BARLINE_SETTINGS = "barline-settings"
+
+    class BhpID:
+        # pylint: disable=too-few-public-methods
+        GROUP_NAME = "BHP Analyzer"
+
+        LINE_CHART = "line-chart"
+
+        FAN_CHART = "fan-chart"
+
+        BAR_CHART = "bar-chart"
