+++ webviz_subsurface/plugins/_bhp_qc/shared_settings/__init__.py
@@ -1,0 +1,1 @@
+from ._filter import BarLineSettings, Filter
