+++ webviz_subsurface/plugins/_bhp_qc/view_elements/__init__.py
@@ -1,0 +1,1 @@
+from ._graph import Graph
