+++ webviz_subsurface/plugins/_bhp_qc/views/__init__.py
@@ -1,0 +1,3 @@
+from ._bar_chart import BarView
+from ._fan_chart import FanView
+from ._line_chart import LineView
